feat(web): move webhook signing secrets to its own /webhook-secrets route
Previously buried as one of six sections inside /settings — promoting
it to a top-level nav entry next to /api-keys. Both are credential
surfaces; surfacing them at the same sidebar depth makes them easier
to find and matches the user's mental model.

  - New /webhook-secrets page owns the create/reveal/rotate/delete
    flow with the same components (banner, table, row) lifted out
    of settings/page.tsx
  - Sidebar gains a new "Webhook secrets" entry between API keys
    and the Settings button. New `shield` icon (shield + checkmark
    silhouette) added to the ICONS set — visually paired with the
    `key` icon so the credential pair reads as related.
  - settings/page.tsx loses the SigningSecretsSection + its 4 helper
    components + the SigningSecretSummary/CreatedSecret types
    (~560 lines removed). Subtitle simplified accordingly.
  - The Jest signing-secrets describe block (~180 lines) and the
    mock-fetch helpers moved to webhook-secrets/page.test.tsx
    verbatim, just retargeted at WebhookSecretsPage. All 8
    signing-secret tests still pass against the new render
    target.

161/161 tests pass (12 suites). Lint clean (1 pre-existing
warning, unrelated). Build green.
